refactor(js): split dashboard cards out of pipeline.js (#434)
pipeline.js was ~1,200 lines mixing the queue list, server-side search,
four dashboard card families, match-rate charts, and item rendering.
The dashboard block (renderPipelineDashboard + cards + charts +
formatters, 570 lines) moves to pipeline_dashboard.js; the queue module
owns the nav strip and passes its HTML in, keeping the dependency
one-way.

tests/test_js_pipeline.mjs splits the same way — and gets ADDED to
scripts/run_tests.sh, where it had never been registered (the JS
pipeline tests only ran when invoked by hand). pipeline_dashboard tests
run alongside.
